add "tensor of tensor" dialect conversion helpers
Adds a utility function `addTensorOfTensorConversionPatterns`` designed to be used as part of dialect conversion.
These support lowerings that would create "tensor of tensor" types, such as what happens to `tensor<..xpoly>` in `--polynomial-to-standard`.

This commit also enables this for `--polynomial-to-standard`, which is a first step towards supporting polynomial operations over tensors in that lowering.